Feat (Module) New\removed confs, automation moderation
So fly for kick and water walking is removed, lets be honest, it was a useless idea.
In its place is the Auto Kick and Auto Ban conf with server wide announcement for kick and ban based on its own unique count conf. How this works is if it reachs a number of reports for either ban or kick, it will automatically enact it. The ban is a account ban.