feat (New Detection) Ignore Z-Axis Check
So this happens when a player is able to block their z-axis from updating with the server. This achieves the ability to run across the sky while avoiding fly hack detection. Usually get on a high object and then run accross, their x and y will chance but their elevation will not.
Warning: At times there will be false positives due to "pot holes" in the vmaps. This can not be avoided unless someone wants to redo all the maps with a better map extractor pr or rework all the maps in recast demo.