Code Reduction
Ebon hold removed due to the teleport helpers being added ina couple of commits ago. Originally the ebon hold provided a aura for DK's that gave a speed boost just in that area that was flagging the speed hack, also the teleporter pad gave a false hit as well for speed hack. both of which no longers happens.