Auto Jailing: Spell Cast Disable, code clean up
Permanent silence casted on offender when using .anticheat jail or if auto jail is turned on, they will not be able to cast any spells to teleport out of there. Code Style Clean up review on the TC Side of the module thanks to @jackpoz that i was able to bring over to this module and remove magic number dealing with  resurrection sickness.