fix (core): Dead Locks
On high player populations deadlocks seem to be caused when you use the .anticheat global gm cmds. SaveAllPlayer is just not necessary or needed in this module which would cause the possible deadlock issues. Its already set to save the player's anticheat data statistics ever minute already. This save all is not need.