update (module): Dereferencing NULL Exploit fix
So there is a "Dereferencing NULL" Exploit that will crash the server thru the anticheat module. That issue is now fixed. I will not be releasing details of that exploit. Just make sure the module is updated and you will not have that problem. Appears to only effect Azerothcore, the TC counter part is not affected.