feat (experimential\conf): Op Code Ack Hack Check
New conf:
Anticheat.OpAckOrderHack = 1

This is experimental. It will run with just the new conf. This checks for the more advance hacks of op code manipulation sent to the server.
It will not alert the GMs and will not add to the collumn counts at the current moment. It will only log into the log file when the conf is enabled.
This is experimental until we get feed back on performance impact.